pyproject.toml
[project]
name = "sql-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai-haystack>=1.3.0",
    "haystack-ai>=2.15.2",
    "pandas>=2.3.1",
    "psycopg>=3.2.9",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "uv>=0.7.19",
]
